+++ app/src/main/java/com/tstreet/onhand/MainActivity.kt
@@ -1,15 +1,12 @@
 package com.tstreet.onhand
 
 import android.os.Bundle
-import android.view.Window
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.compose.runtime.CompositionLocalProvider
 import androidx.core.content.ContextCompat
 import com.tstreet.onhand.core.common.LocalCommonProvider
 import com.tstreet.onhand.core.data.di.LocalDataProvider
-import com.tstreet.onhand.core.ui.theming.Grey10
-import com.tstreet.onhand.core.ui.theming.MATTE_GREEN
 import com.tstreet.onhand.core.ui.theming.OnHandTheme
 import com.tstreet.onhand.nav.Navigation
