+++ app/src/main/java/com/tstreet/onhand/nav/Navigation.kt
@@ -17,8 +17,8 @@ import com.tstreet.onhand.core.common.LocalCommonProvider
 import com.tstreet.onhand.core.common.injectedViewModel
 import com.tstreet.onhand.core.data.di.LocalDataProvider
 import com.tstreet.onhand.core.ui.RECIPE_ID_NAV_KEY
-import com.tstreet.onhand.feature.ingredientsearch.IngredientSearchScreen
-import com.tstreet.onhand.feature.ingredientsearch.di.DaggerIngredientSearchComponent
+import com.tstreet.onhand.feature.home.HomeScreen
+import com.tstreet.onhand.feature.home.di.DaggerHomeComponent
 import com.tstreet.onhand.feature.recipedetail.INVALID_RECIPE_ID
 import com.tstreet.onhand.feature.recipedetail.RecipeDetailScreen
 import com.tstreet.onhand.feature.recipedetail.di.DaggerRecipeDetailComponent
@@ -64,9 +64,9 @@ private fun NavigationConfiguration(
         // Note: each composable { } block is triggered for each recomposition (potentially as
         // often each new frame). Revisit whether this is a performance issue later.
         composable(route = BottomNavigationScreen.IngredientSearch.route) {
-            IngredientSearchScreen(
+            HomeScreen(
                 injectedViewModel {
-                    DaggerIngredientSearchComponent
+                    DaggerHomeComponent
                         .builder()
                         .dataComponentProvider(dataProvider)
                         .commonComponentProvider(commonProvider)
