+++ feature/home/src/test/java/com/tstreet/onhand/feature/home/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package com.tstreet.onhand.feature.ingredientsearch
+package com.tstreet.onhand.feature.home
 
 import org.junit.Test
