--- feature/ingredientsearch/src/main/java/com/tstreet/onhand/feature/ingredientsearch/di/IngredientSearchScope.kt
@@ -1,8 +1,0 @@
-package com.tstreet.onhand.feature.ingredientsearch.di
-
-import javax.inject.Scope
-
-@Scope
-@Retention(AnnotationRetention.RUNTIME)
-@Deprecated("Potentially use @FeatureScope instead, see doc for details.")
-annotation class IngredientSearchScope
