+++ settings.gradle
@@ -24,7 +24,7 @@ include ':core:network'
 include ':core:common'
 include ':core:domain'
 include ':core:database'
-include ':feature:ingredientsearch'
+include ':feature:home'
 include ':feature:shoppinglist'
 include ':feature:recipesearch'
 include ':feature:recipedetail'
